requirements.txt
# Django itself
Django>=3.0,<4.0

# Additional Django Apps
django-forms-bootstrap>=3.1.0

# For highlighting
Markdown==3.3.4

# For filters
pre-js-py==1.1.0

# For testing pep
pep8 >= 1.7.0

# For authing against dreamjub
django-allauth==0.29.0
django-allauth-dreamjub==0.1.3

# for counting users
requests-oauthlib==0.8.0
